Make the global list of testExpressions local
